Merge: Fix a regression: not exit after saving a modified file

+++ PowerEditor/src/Notepad_plus.cpp
@@ -7190,7 +7190,7 @@ void Notepad_plus::restoreMinimizeDialogs()
 {
 	_dockingManager.showFloatingContainers(true);
 	size_t nbDialogs = _sysTrayHiddenHwnd.size();
-	for (int i = (nbDialogs - 1); i >= 0; i--)
+	for (int i = (static_cast<int>(nbDialogs) - 1); i >= 0; i--)
 	{
 		::ShowWindow(_sysTrayHiddenHwnd[i], SW_SHOW);
 		_sysTrayHiddenHwnd.erase(_sysTrayHiddenHwnd.begin() + i);
